nebezdariproject/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-appconf==1.0.4
django-ckeditor==5.9.0
django-extensions==2.2.9
django-filter==2.3.0
django-js-asset==1.2.2
django-select2==7.2.3
djangorestframework==3.11.0
djangorestframework-filters==1.0.0.dev0
drf-extensions==0.6.0
drf-yasg==1.17.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.4
Pillow==7.1.1
pydot==1.4.1
pyparsing==2.4.7
pyrecord==1.0.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
